requirements.txt
openai>=1.12.0
chromadb>=0.4.22
pypdf>=4.0.0
requests>=2.31.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
nltk>=3.8.0
feedparser>=6.0.10
sentence-transformers>=2.2.0
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
